pyproject.toml
[project]
name = "stock-volatility-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "feedparser>=6.0.12",
    "flask>=3.1.2",
    "google-genai>=1.59.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.4",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.6",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.40.0",
    "yfinance>=1.0",
]
